requirements.txt
numpy==1.23.3
pandas==1.5.0
pandas_datareader==0.10.0
Pillow==9.2.0
streamlit==1.13.0
